requirements.txt
amqp==2.2.1
billiard==3.5.0.3
celery==4.1.0
defusedxml==0.5.0
Django==1.11.4
django-filter==1.0.4
django-mptt==0.8.7
django-rest-framework==0.1.0
django-simple-history==1.9.0
django-websocket-redis==0.5.0
djangorestframework==3.6.3
djangorestframework-xml==1.3.0
Faker==0.7.18
gevent==1.2.2
greenlet==0.4.12
kombu==4.1.0
psycopg2==2.7.3
python-dateutil==2.6.1
pytz==2017.2
redis==2.10.5
six==1.10.0
uWSGI==2.0.15
vine==1.1.4
